refactor: convert controlUtils to TypeScript
